testing: refactored internal/clock to only support injection when 'testing' build tag is present
