+++ internal/clock/now.go
@@ -2,18 +2,8 @@
 package clock
 
 import (
-	"encoding/json"
-	"log"
-	"net/http"
-	"os"
-	"sync"
 	"time"
 )
-
-const refreshServerTimeEvery = 3 * time.Second
-
-// Now is overridable function that returns current wall clock time.
-var Now = time.Now // allow:no-inject-time
 
 // Since returns time since the given timestamp.
 func Since(t time.Time) time.Duration {
@@ -23,61 +13,4 @@ func Since(t time.Time) time.Duration {
 // Until returns duration of time between now and a given time.
 func Until(t time.Time) time.Duration {
 	return t.Sub(Now())
-}
-
-func init() {
-	fakeTimeServer := os.Getenv("KOPIA_FAKE_CLOCK_ENDPOINT")
-	if fakeTimeServer == "" {
-		return
-	}
-
-	Now = getTimeFromServer(fakeTimeServer)
-}
-
-// getTimeFromServer returns a function that will return timestamp as returned by the server
-// increasing it client-side by certain inteval until maximum is reached, at which point
-// it will ask the server again for new timestamp.
-//
-// The server endpoint must be HTTP and be set using KOPIA_FAKE_CLOCK_ENDPOINT environment
-// variable.
-func getTimeFromServer(endpoint string) func() time.Time {
-	var mu sync.Mutex
-
-	var timeInfo struct {
-		Next  int64 `json:"next"`
-		Step  int64 `json:"step"`
-		Until int64 `json:"until"`
-	}
-
-	nextRefreshRealTime := time.Now() // allow:no-inject-time
-
-	return func() time.Time {
-		mu.Lock()
-		defer mu.Unlock()
-
-		if timeInfo.Next >= timeInfo.Until || time.Now().After(nextRefreshRealTime) { // allow:no-inject-time
-			resp, err := http.Get(endpoint) //nolint:gosec,noctx
-			if err != nil {
-				log.Fatalf("unable to get fake time from server: %v", err)
-			}
-			defer resp.Body.Close() //nolint:errcheck
-
-			if resp.StatusCode != http.StatusOK {
-				log.Fatalf("unable to get fake time from server: %v", resp.Status)
-			}
-
-			if err := json.NewDecoder(resp.Body).Decode(&timeInfo); err != nil {
-				log.Fatalf("invalid time received from fake time server: %v", err)
-			}
-
-			nextRefreshRealTime = time.Now().Add(refreshServerTimeEvery) // allow:no-inject-time
-		}
-
-		v := timeInfo.Next
-		timeInfo.Next += timeInfo.Step
-
-		n := time.Unix(0, v)
-
-		return n
-	}
 }
