+++ internal/clock/now_prod.go
@@ -1,0 +1,10 @@
+// +build !testing
+
+package clock
+
+import "time"
+
+// Now returns current wall clock time.
+func Now() time.Time {
+	return time.Now() // allow:no-inject-time
+}
